riscv: dts: fu540: mark CPU 0 as disabled for Linux
On the SiFive FU540, CPU 0 is an MMU-less "microcontroller" that isn't
intended to run the Linux kernel.  For this reason, most versions of
BBL, and apparently OpenSBI, edit the DT to mask this CPU off before
passing it to Linux.  However, we wish to minimize the number of DT
modifications made by early boot components. aside from board-specific
device discovery tasks such as filling in the amount of memory on the
board, and chosen node changes.  We therefore mark CPU 0 as disabled
in the Linux kernel DT (and drop the corresponding DT modification code
in BBL).