[NOT FOR UPSTREAMING] firmware: psci: add an experimental SBI conduit for RISC-V
Prototype a RISC-V specific SBI conduit for PSCI.  This uses the SBI
calling convention.  The goal here is to reuse as much of PSCI as
possible for RISC-V.  This is primarily targeted towards the system
management operations like system reset, shutdown, suspend, etc.; but
should also be extensible to CPU up/down operations, migration, etc.

This work is facilitated by the existing PSCI firmware code's support
for multiple conduits for ARM.  This implementation is a bit rough,
but should be easy to refactor.

Compile-tested only.

This is an experimental patch; not for upstreaming.